refactor(app.admin): share + enlarge split-pane layout, refine breakpoints
The split-pane styles were duplicated across the four entity pages. Extract
them into a single shared module (components/detail/splitPane.css.ts) that
each page CSS re-exports, so the layout can't drift and is tuned in one place.

Tuning:
- Detail pane is larger and fluid: flex-basis clamp(440px, 44%, 760px) so it
  grows with the viewport up to a generous cap instead of a fixed 440px / 500px
  max. The list keeps a 320px readable minimum on desktop.
- Single-column collapse documented at the ≤1024px (tablet/mobile) breakpoint;
  the "Back to list" control gains a 44px minimum touch target for mobile.
